consolidate labware fixtures to one dir
