+++ components/src/deck/LabwareRender.md
@@ -3,10 +3,10 @@ Note that `LabwareRender` is in robot coordinates, we transform it to view it us
 **Normal Labware**
 
 ```js
-const fixture96Plate = require('@opentrons/shared-data/fixtures/fixture96Plate')
-const fixture24TubeRack = require('@opentrons/shared-data/fixtures/fixture24TubeRack')
-const fixture12Trough = require('@opentrons/shared-data/fixtures/fixture12Trough')
-const fixtureTipRack300Ul = require('@opentrons/shared-data/fixtures/fixtureTipRack300Ul')
+const fixture96Plate = require('@opentrons/shared-data/labware/fixtures/2/fixture96Plate')
+const fixture24TubeRack = require('@opentrons/shared-data/labware/fixtures/2/fixture24TubeRack')
+const fixture12Trough = require('@opentrons/shared-data/labware/fixtures/2/fixture12Trough')
+const fixtureTipRack300Ul = require('@opentrons/shared-data/labware/fixtures/2/fixtureTipRack300Ul')
 
 // Change this to view different labware fixtures
 let definition = fixture96Plate
