+++ protocol-designer/src/file-data/__tests__/commandsSelectors.test.js
@@ -1,8 +1,7 @@
-import {
-  fixture12Trough,
-  fixture96Plate,
-  fixtureTrash,
-} from '@opentrons/shared-data/fixtures'
+import fixture12Trough from '@opentrons/shared-data/labware/fixtures/2/fixture12Trough.json'
+import fixture96Plate from '@opentrons/shared-data/labware/fixtures/2/fixture96Plate.json'
+import fixtureTrash from '@opentrons/shared-data/labware/fixtures/2/fixtureTrash.json'
+
 import { getLabwareLiquidState } from '../selectors'
 jest.mock('../../labware-defs/utils')
