+++ protocol-designer/src/step-generation/test-with-flow/robotStateSelectors.test.js
@@ -1,5 +1,6 @@
 // @flow
-import { fixtureTipRack300Ul } from '@opentrons/shared-data/fixtures'
+import fixtureTipRack300Ul from '@opentrons/shared-data/labware/fixtures/2/fixtureTipRack300Ul.json'
+
 import {
   makeContext,
   makeState,
