+++ protocol-designer/src/well-selection/test/utils.test.js
@@ -1,8 +1,6 @@
-import {
-  fixture12Trough,
-  fixture96Plate,
-  fixture384Plate,
-} from '@opentrons/shared-data/fixtures'
+import fixture12Trough from '@opentrons/shared-data/labware/fixtures/2/fixture12Trough.json'
+import fixture96Plate from '@opentrons/shared-data/labware/fixtures/2/fixture96Plate.json'
+import fixture384Plate from '@opentrons/shared-data/labware/fixtures/2/fixture384Plate.json'
 import { getWellSetForMultichannel } from '../utils'
 
 describe('getWellSetForMultichannel (integration test)', () => {
