+++ shared-data/js/__tests__/definition2-schema.test.js
@@ -2,8 +2,8 @@ import path from 'path'
 import glob from 'glob'
 import Ajv from 'ajv'
 import schema from '../../labware-json-schema/labwareSchemaV2.json'
-import exampleLabware1 from './fixtures/labware2/labwareExample.json'
-import exampleLabware2 from './fixtures/labware2/labwareExample2.json'
+import exampleLabware1 from '../../labware/fixtures/2/labwareExample1.json'
+import exampleLabware2 from '../../labware/fixtures/2/labwareExample2.json'
 
 const definitionsGlobPath = path.join(
   __dirname,
